Move navigation's NavView into include/, src/ files
