+++ main.cpp
@@ -1,5 +1,3 @@
-#define ALLEGRO_UNSTABLE
-
 #include <allegro_flare/allegro_flare.h>
 #include <cmath>
 
@@ -7,7 +5,7 @@
 
 #include <style_assets.hpp>
 #include <logging.hpp>
-#include "target_id.hpp"
+#include <target_id.hpp>
 #include <inventory_gui/screen.hpp>
 #include "start_screen.hpp"
 #include "world_render.hpp"
